fix: recover CI and persist claims atomically (#789)

+++ tests/support/mod.rs
@@ -90,7 +90,7 @@ fn free_space_bytes(path: &Path) -> Option<u64> {
     if unsafe { libc::statvfs(path_c.as_ptr(), &mut stat) } != 0 {
         return None;
     }
-    Some(stat.f_bavail.saturating_mul(stat.f_frsize) as u64)
+    Some((stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64))
 }
 
 #[cfg(not(unix))]
